moved conformerblock.
